+++ internal/query/client.go
@@ -41,9 +41,9 @@ type (
 		With(ctx context.Context, f func(ctx context.Context, s *Session) error, opts ...retry.Option) error
 	}
 	Client struct {
-		config *config.Config
-		client Ydb_Query_V1.QueryServiceClient
-		pool   sessionPool
+		config              *config.Config
+		client              Ydb_Query_V1.QueryServiceClient
+		explicitSessionPool sessionPool
 
 		// implicitSessionPool is a pool of implicit sessions,
 		// i.e. fake sessions created without CreateSession/AttachSession requests.
@@ -198,7 +198,7 @@ func (c *Client) Close(ctx context.Context) error {
 
 	close(c.done)
 
-	if err := c.pool.Close(ctx); err != nil {
+	if err := c.explicitSessionPool.Close(ctx); err != nil {
 		return xerrors.WithStackTrace(err)
 	}
 
@@ -246,7 +246,7 @@ func (c *Client) Do(ctx context.Context, op query.Operation, opts ...options.DoO
 		onDone(attempts, finalErr)
 	}()
 
-	err := do(ctx, c.pool,
+	err := do(ctx, c.explicitSessionPool,
 		func(ctx context.Context, s *Session) error {
 			return op(ctx, s)
 		},
@@ -333,12 +333,7 @@ func (c *Client) QueryRow(ctx context.Context, q string, opts ...options.Execute
 		onDone(finalErr)
 	}()
 
-	pool := c.pool
-	if c.config.IsImplicitSession() {
-		pool = c.implicitSessionPool
-	}
-
-	row, err := clientQueryRow(ctx, pool, q, settings, withTrace(c.config.Trace()))
+	row, err := clientQueryRow(ctx, c.pool(), q, settings, withTrace(c.config.Trace()))
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -385,12 +380,7 @@ func (c *Client) Exec(ctx context.Context, q string, opts ...options.Execute) (f
 		onDone(finalErr)
 	}()
 
-	pool := c.pool
-	if c.config.IsImplicitSession() {
-		pool = c.implicitSessionPool
-	}
-
-	err := clientExec(ctx, pool, q, opts...)
+	err := clientExec(ctx, c.pool(), q, opts...)
 	if err != nil {
 		return xerrors.WithStackTrace(err)
 	}
@@ -438,12 +428,7 @@ func (c *Client) Query(ctx context.Context, q string, opts ...options.Execute) (
 		onDone(err)
 	}()
 
-	pool := c.pool
-	if c.config.IsImplicitSession() {
-		pool = c.implicitSessionPool
-	}
-
-	r, err = clientQuery(ctx, pool, q, opts...)
+	r, err = clientQuery(ctx, c.pool(), q, opts...)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -498,17 +483,23 @@ func (c *Client) QueryResultSet(
 		onDone(finalErr, rowsCount)
 	}()
 
-	pool := c.pool
-	if c.config.IsImplicitSession() {
-		pool = c.implicitSessionPool
-	}
-
-	rs, rowsCount, err = clientQueryResultSet(ctx, pool, q, settings, withTrace(c.config.Trace()))
+	rs, rowsCount, err = clientQueryResultSet(ctx, c.pool(), q, settings, withTrace(c.config.Trace()))
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
 
 	return rs, nil
+}
+
+// pool returns the appropriate session pool based on the client configuration.
+// If implicit sessions are enabled, it returns the implicit session pool;
+// otherwise, it returns the explicit session pool.
+func (c *Client) pool() sessionPool {
+	if c.config.AllowImplicitSessions() {
+		return c.implicitSessionPool
+	}
+
+	return c.explicitSessionPool
 }
 
 func (c *Client) DoTx(ctx context.Context, op query.TxOperation, opts ...options.DoTxOption) (finalErr error) {
@@ -527,7 +518,7 @@ func (c *Client) DoTx(ctx context.Context, op query.TxOperation, opts ...options
 		onDone(attempts, finalErr)
 	}()
 
-	err := doTx(ctx, c.pool, op,
+	err := doTx(ctx, c.explicitSessionPool, op,
 		settings.TxSettings(),
 		append(
 			[]retry.Option{
@@ -594,7 +585,7 @@ func New(ctx context.Context, cc grpc.ClientConnInterface, cfg *config.Config) *
 		client:              client,
 		done:                make(chan struct{}),
 		implicitSessionPool: createImplicitSessionPool(ctx, cfg, client, cc),
-		pool: pool.New(ctx,
+		explicitSessionPool: pool.New(ctx,
 			pool.WithLimit[*Session](cfg.PoolLimit()),
 			pool.WithItemUsageLimit[*Session](cfg.PoolSessionUsageLimit()),
 			pool.WithItemUsageTTL[*Session](cfg.PoolSessionUsageTTL()),
@@ -649,6 +640,7 @@ func createImplicitSessionPool(ctx context.Context,
 ) sessionPool {
 	return pool.New(ctx,
 		pool.WithLimit[*Session](cfg.PoolLimit()),
+		pool.WithTrace[*Session](poolTrace(cfg.Trace())),
 		pool.WithCreateItemFunc(func(ctx context.Context) (_ *Session, err error) {
 			core := &sessionCore{
 				cc:     cc,
@@ -656,8 +648,6 @@ func createImplicitSessionPool(ctx context.Context,
 				Trace:  cfg.Trace(),
 				done:   make(chan struct{}),
 			}
-
-			core.SetStatus(StatusIdle)
 
 			return &Session{
 				Core:   core,