+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-* Added a feature: "Execute queries without explicit session creation":
-  * added `config.WithImplicitSessions` query option for execute queries with implicit sessions
+* Added `query.AllowImplicitSessions()` option for execute queries through `query.Client.{Exec,Query,QueryResultSet,QueryRow}` without explicit sessions
 
 ## v3.112.0
 * Added support for the `json.Unmarshaler` interface in the `CastTo` function for use in scanners, such as the `ScanStruct` method